+++ CHANGES.md
@@ -1,4 +1,8 @@
 # Release History
 
+## 0.2.0
+- use `func_get_args()` now to support variable length function arguments,
+  no longer need to provide an array
+
 ## 0.1.0
 - initial release